Start moving from UI to headless mode.

+++ src/main/kotlin/StockSymbol.kt
@@ -2,7 +2,7 @@ import tornadofx.getProperty
 import tornadofx.property
 import java.util.*
 
-open class StockSymbol(date: Date, open: Float, high: Float, low: Float, close: Float, volume: Float, adjClose: Float) {
+open class StockSymbol(date: Date, open: Float, high: Float, low: Float, close: Float, volume: Int, adjClose: Float) {
     var date: Date by property(date)
     fun dateProperty() = getProperty(StockSymbol::date)
 
@@ -18,7 +18,7 @@ open class StockSymbol(date: Date, open: Float, high: Float, low: Float, close: 
     var close: Float by property(close)
     fun closeProperty() = getProperty(StockSymbol::close)
 
-    var volume: Float by property(volume)
+    var volume: Int by property(volume)
     fun volumeProperty() = getProperty(StockSymbol::volume)
 
     var adjClose: Float by property(adjClose)
